requirements.txt
absl-py==0.13.0
astroid==2.6.2
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.1
cloudpickle==1.6.0
dartpy==0.0.1.post11
# dataclasses==0.8
decorator==4.4.2
dm-sonnet==2.0.0
dm-tree==0.1.6
flatbuffers==1.12
freetype-py==2.2.0
future==0.18.2
gast==0.3.3
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graph-nets==1.1.0
grpcio==1.32.0
h5py==2.10.0
idna==3.2
imageio==2.9.0
importlib-metadata==4.6.1
isort==5.9.2
Keras-Preprocessing==1.1.2
lazy-object-proxy==1.6.0
lxml==4.6.3
Markdown==3.3.4
mccabe==0.6.1
networkx==2.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
Pillow==8.3.1
pkg-resources==0.0.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.6
PyContracts==1.8.12
PyGeometry==1.5.6
pyglet==1.5.18
pylint==2.9.3
PyOpenGL==3.1.0
pyparsing==2.4.7
pyrender==0.1.45
python-dateutil==2.8.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-probability==0.13.0
termcolor==1.1.0
toml==0.10.2
trimesh==3.9.21
typed-ast==1.4.3
typing-extensions==3.7.4.3
urdfpy==0.0.22
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
